Allow the user to specify a tmpdir for hooks
It is common to mount /tmp as noexec on linux, but pkgr uses mktemp to
create a tmpfile in which to output the custom pre/post
install/uninstall scripts, and then tries to exec them. This means that
pkgr generates RPMs that don't function correctly on systems with /tmp
mounted as noexec.

For lots of reasons (yum update being the main one), it is not ideal to
always have the caller set TMPDIR before installing the package, so this
exposes a new command line option (--tmpdir) that allows the user to
specify an alternate directory in which to create the tmp files.